Vendor rbzip2 for now because the latest release is 2 years behind master.
This is heavily modified to remove extra code, including the FFI and JRuby implementations in favor of slimmer code. I also removed the compressor because I'm not using it. The Decompressor has been slightly modified to work with TarReader.

Given that RBzip2 hasn't had a release in 3 years, it seems unlikely I'll ever need to update this anyway.